+++ src/components/absolute/overlays.js
@@ -1,4 +1,4 @@
 export default {
   start: 1,
-  end: 2,  
+  end: 2,
 };